+++ internal/translation/client.go
@@ -1,4 +1,4 @@
-package translayer
+package translation
 
 import (
 	"context"
